requirements.txt
torch
torchvision
torchaudio
pandas
matplotlib
seaborn
tensorboardX
ipykernel
ipywidgets<8.0.0
tqdm
scikit-learn
pytest
rootutils
pytorch-lightning
hydra-core
x-transformers
uszipcode
plotly
intervaltree
black
isort
pytest-cov
ray[tune]
xgboost
xformers